requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
aiosqlite==0.19.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
factory-boy==3.3.0
pytest-cov==4.1.0

# Development tools
black==23.11.0
isort==5.12.0
mypy==1.7.1
pre-commit==3.6.0

# File handling
aiofiles==23.2.1
python-magic==0.4.27

# Environment
python-dotenv==1.0.0
